Update CREDITS.md with new links for me
Also updated the Phoenix Gaming Network link to the Frag Haven link (we're in the process of switching the community to a new name).


Hope this is okay. Also would like to be put back into the contributors section (I haven't made any since the switch from Git.tf to GitHub, though I plan to take on some of the bigger tasks/issues soon).

Sorry for my absence guys.